requirements.txt
numpy==1.21.4
openpyxl==3.0.9
pandas==1.3.4
protobuf==3.19.0
psycopg2-binary==2.9.3
pytz==2021.3
requests==2.26.0
SQLAlchemy==1.4.29
streamlit==1.10.0
boto3==1.18.7
streamlit-aggrid==0.3.2

